API authentication finished, controllers fixed, responses fixed

+++ src/main/java/aiss/videominer/controller/ChannelController.java
@@ -16,7 +16,7 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/videominer/channels")
+@RequestMapping("/videominer/v1")
 public class ChannelController {
 
     @Autowired
@@ -37,29 +37,49 @@ public class ChannelController {
     @Autowired
     TokenRepository tokenRepository;
 
-    // GET http://localhost:8080/videominer/channels
-    @GetMapping
-    public List<Channel> findAll() { return channelRepository.findAll();}
-
-    // GET http://localhost:8080/videominer/channels/{id}
-    @GetMapping("/{id}")
-    public Channel findById(@PathVariable String id) throws ChannelNotFoundException {
-        Optional<Channel> channel = channelRepository.findById(id);
-        if (!channel.isPresent()) {
-            throw new ChannelNotFoundException();
+    // GET http://localhost:8080/videominer/v1/channels
+    @ResponseStatus(HttpStatus.OK)
+    @GetMapping("/channels")
+    public List<Channel> findAll(@RequestHeader HttpHeaders header) throws TokenNotValidException, TokenRequiredException {
+        String token = header.getFirst("Authorization");
+        if (token==null) {
+            throw new TokenRequiredException();
         }
-        return channel.get();
+        else if(tokenRepository.existsById(token)) {
+            return channelRepository.findAll();
+        } else {
+            throw new TokenNotValidException();
+        }
     }
 
-    //POST http://localhost:8080/videominer/channels
+    // GET http://localhost:8080/videominer/v1/channels/{id}
+    @ResponseStatus(HttpStatus.OK)
+    @GetMapping("/channels/{id}")
+    public Channel findById(@PathVariable String id, @RequestHeader HttpHeaders header) throws ChannelNotFoundException, TokenRequiredException, TokenNotValidException {
+        String token = header.getFirst("Authorization");
+        if (token==null) {
+            throw new TokenRequiredException();
+        }
+        else if(tokenRepository.existsById(token)) {
+            Optional<Channel> channel = channelRepository.findById(id);
+            if (!channel.isPresent()) {
+                throw new ChannelNotFoundException();
+            }
+            return channel.get();
+        } else {
+            throw new TokenNotValidException();
+        }
+    }
+
+    //POST http://localhost:8080/videominer/v1/channels
     @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping
+    @PostMapping("/channels")
     public Channel create(@Valid @RequestBody Channel channel, @RequestHeader HttpHeaders header) throws TokenNotValidException, TokenRequiredException {
         String token = header.getFirst("Authorization");
         if (token==null) {
             throw new TokenRequiredException();
         }
-        else if(tokenRepository.findById(token).isPresent()) {
+        else if(tokenRepository.existsById(token)) {
             Channel _channel = channelRepository.save(channel);
             for (Video v : channel.getVideos()) {
                 Video video = videoRepository.save(v);
@@ -77,25 +97,45 @@ public class ChannelController {
         }
     }
 
-    // PUT http://localhost:8080/videominer/channels/{id}
-    @PutMapping("/{id}")
-    public void update(@Valid @RequestBody Channel updatedChannel, @PathVariable String id) throws ChannelNotFoundException {
-        Optional<Channel> channelData = channelRepository.findById(id);
-        if (!channelData.isPresent()) {
-            throw new ChannelNotFoundException();
+    // PUT http://localhost:8080/videominer/v1/channels/{id}
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PutMapping("/channels/{id}")
+    public void update(@Valid @RequestBody Channel updatedChannel, @PathVariable String id, @RequestHeader HttpHeaders header) throws ChannelNotFoundException, TokenRequiredException, TokenNotValidException {
+        String token = header.getFirst("Authorization");
+        if (token==null) {
+            throw new TokenRequiredException();
         }
-        Channel _channel = channelData.get();
-        _channel.setName(updatedChannel.getName());
-        _channel.setDescription(updatedChannel.getDescription());
-        channelRepository.save(_channel);
+        else if(tokenRepository.existsById(token)) {
+            Optional<Channel> channelData = channelRepository.findById(id);
+            if (!channelData.isPresent()) {
+                throw new ChannelNotFoundException();
+            }
+            Channel _channel = channelData.get();
+            _channel.setName(updatedChannel.getName());
+            _channel.setDescription(updatedChannel.getDescription());
+            channelRepository.save(_channel);
+        } else {
+            throw new TokenNotValidException();
+        }
     }
 
-    // DELETE http://localhost:8080/videominer/channels/{id}
+    // DELETE http://localhost:8080/videominer/v1/channels/{id}
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    @DeleteMapping("/{id}")
-    public void delete(@PathVariable String id) {
-        if(channelRepository.existsById(id)) {
-            channelRepository.deleteById(id);
+    @DeleteMapping("/channels/{id}")
+    public void delete(@PathVariable String id, @RequestHeader HttpHeaders header) throws TokenRequiredException, TokenNotValidException, ChannelNotFoundException {
+        String token = header.getFirst("Authorization");
+        if (token==null) {
+            throw new TokenRequiredException();
+        }
+        else if(tokenRepository.existsById(token)) {
+            if(channelRepository.existsById(id)) {
+                channelRepository.deleteById(id);
+            }
+            else{
+                throw new ChannelNotFoundException();
+            }
+        } else {
+            throw new TokenNotValidException();
         }
     }
 }
